CrossWord now with 2 intersecting words
- Managed to get "DINOSAUR" and "RABBIT" to intersect making serveral
initial checks of the following:
1. Finding any intersections
2. If finding possible intersection, check RIGHT, LEFT and TOP, BOTTOM
to see whether pairs are filled or not. Note: Both either Right or Left/
Top or Bottom must be empty for across/bottom to be considered.
3. When and if across or down is confirmed, check to see whether there
is enough space for the letters of the word to placed on either side of
the intersections
4. If step 3 is met, check to see whether these places have been filled
in (aside from place of intersection). If empty slots, place remaining
letters in relevant spots.

--HAVE INCLUDED DEBUGGING FILE FOR TESTING, CAN BE USED LATER IN REPORT
MAYBE?