Image moving across screen
- Managed to get image moving according to changes in either X or Y or Z
however does not correspond the way I want it to therefore changing over
to new algorithm as described below....

- NEW ALGORITHM:

1. Use Bearing instead of X, Y, Z values of accelerometer
2. Randomly select a bearing for an object
3. If bearing of device matches the bearing of item, it will be
displayed bang in the middle of screen
3. If bearing of object is less than current bearing of device, object
will be displayed to the right. And vice versa, if bearing of object is
more, it will be displayed to the left (Bearings from North increase in
clockwise rotation)

4. Distance in meters will be a multiplier. The furthur the time, the
less it moves on either side and the closer the item, the more it moves
either side. To test this, use a reference box, like a black piece of
paper to see whether it stays in the same spot just by moving closer of
farther away (GOOD FOR TESTING SECTION).

CURRENT PROBLEM:
- Bearing is a bit erratic when device is lifted off surface and held in
a vertical position...